+++ backend/controllers/user.controller.js
@@ -49,11 +49,13 @@ const  deleteUser = async (req, res) => {
   const {id} = req.params;
 
   try {
+    if(req.payload.role === "admin") {
     const User = await User.findByIdAndDelete(id);
 
     if(!User) return res.status(404).json({message: "User not found"});
 
     return res.status(200).json({message: "Delete successfully!"});
+    } else return res.status(403).json({message: "Only admin can delete user!"})
   } catch (error) {
     return res.status(500).json({message: error});
   }
